requirements.txt
langchain
langchain-core
openai
chromadb
PyPDF2
tqdm
sentence-transformers
gradio
python-dotenv
